refactor(ai): 双 AI 应用统一调度重构 — stream_run 统一分派 + skill MECE 收敛 + 报告三步流水线

+++ server/apps/backend/app/models/valuation.py
@@ -1,6 +1,7 @@
 from datetime import datetime
+from decimal import Decimal
 
-from sqlalchemy import BigInteger, DateTime, Float, ForeignKey, Text, func
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Numeric, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from apps.backend.app.database import Base
@@ -12,6 +13,6 @@ class AssetValuation(Base):
 
     id: Mapped[int] = mapped_column(BigInteger, primary_key=True, default=next_id)
     asset_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("assets.id"), nullable=False)
-    value: Mapped[float] = mapped_column(Float, nullable=False)
+    value: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
     valued_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
